requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.6
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.4
packaging==23.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
PyJWT==2.8.0
pyTelegramBotAPI==4.14.0
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.3
